Remove nav header; move help and logout btns to bottom of new side nav
