+++ clients/admin-ui/src/features/common/nav/v2/MainSideNav.tsx
@@ -8,14 +8,26 @@ import {
   Button,
   Link,
   ListItem,
+  Menu,
+  MenuButton,
+  MenuDivider,
+  MenuItem,
+  MenuList,
+  QuestionIcon,
+  Stack,
   UnorderedList,
+  UserIcon,
   VStack,
+  Text,
 } from "@fidesui/react";
 import NextLink from "next/link";
 import { useRouter } from "next/router";
 
 import logoImage from "~/../public/logo-white.svg";
+import { useAppDispatch, useAppSelector } from "~/app/hooks";
+import { logout, selectUser, useLogoutMutation } from "~/features/auth";
 import Image from "~/features/common/Image";
+import { User } from "~/features/user-management/types";
 
 import { useNav } from "./hooks";
 import { ActiveNav, NavGroup, NavGroupChild } from "./nav-config";
@@ -120,34 +132,82 @@ const NavGroupMenu = ({
 export const UnconnectedMainSideNav = ({
   groups,
   active,
+  handleLogout,
+  username
 }: {
   groups: NavGroup[];
   active: ActiveNav | undefined;
+  handleLogout: any;
+  username: User;
 }) => (
-  <Box p={4} minWidth="200px" maxWidth="200px" backgroundColor="#191D27">
-    <VStack as="nav" alignItems="start" color="white" height="100%">
-      <Box pb={6}>
-        <FidesLogoHomeLink />
+  <Box p={4} minWidth="200px" maxWidth="200px" backgroundColor="#191D27" height={"100%"} overflow={"scroll"}>
+    <VStack as="nav" alignItems="start" color="white" height={"100%"} justifyContent={"space-between"}>
+      <Box>
+        <Box pb={6}>
+          <FidesLogoHomeLink />
+        </Box>
+        <Accordion
+          allowMultiple
+          width="100%"
+          defaultIndex={[...Array(groups.length).keys()]}
+          overflowY="auto"
+        >
+          {groups.map((group) => (
+            <NavGroupMenu key={group.title} group={group} active={active} />
+          ))}
+        </Accordion>
       </Box>
-      <Accordion
-        allowMultiple
-        width="100%"
-        defaultIndex={[...Array(groups.length).keys()]}
-        overflowY="auto"
-      >
-        {groups.map((group) => (
-          <NavGroupMenu key={group.title} group={group} active={active} />
-        ))}
-      </Accordion>
+      <Box alignItems="center">
+        <Link href="https://docs.ethyca.com" isExternal>
+          <Button size="sm" variant="ghost">
+            <QuestionIcon color="white" boxSize={4} />
+          </Button>
+        </Link>
+        {username && (
+          <Menu>
+            <MenuButton
+              as={Button}
+              size="sm"
+              variant="ghost"
+              data-testid="header-menu-button"
+            >
+              <UserIcon color="white" />
+            </MenuButton>
+            <MenuList shadow="xl" zIndex="20">
+              <Stack px={3} py={2} spacing={1}>
+                <Text color={"gray.700"} fontWeight="medium">{username.username}</Text>
+              </Stack>
+
+              <MenuDivider />
+              <MenuItem
+                color={"gray.700"}
+                _focus={{ color: "complimentary.500", bg: "gray.100" }}
+                onClick={handleLogout}
+                data-testid="header-menu-sign-out"
+              >
+                Sign out
+              </MenuItem>
+            </MenuList>
+          </Menu>
+        )}
+      </Box>
     </VStack>
   </Box>
+
 );
 
 const MainSideNav = () => {
   const router = useRouter();
   const nav = useNav({ path: router.pathname });
+  const [logoutMutation] = useLogoutMutation();
+  const dispatch = useAppDispatch();
+  const username = useAppSelector(selectUser) ?? { username: "" }; // Define username here
 
-  return <UnconnectedMainSideNav {...nav} />;
+  const handleLogout = async () => {
+    await logoutMutation({});
+    dispatch(logout());
+  };
+  return <UnconnectedMainSideNav {...nav} handleLogout={handleLogout} username={username} />;
 };
 
 export default MainSideNav;
